+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for datetime and timestamp function in legacy
+* Support for the Logstash notification system
 
 ### Changed